+++ .eslintrc.js
@@ -5,8 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-const path = require('path');
-const customImportResolver = path.resolve('./eslintImportResolver');
+const customImportResolver = require.resolve('./eslintImportResolver');
 
 module.exports = {
   extends: [
@@ -16,6 +15,20 @@ module.exports = {
     'prettier/flowtype',
   ],
   overrides: [
+    {
+      files: ['*.ts', '*.tsx'],
+      parser: '@typescript-eslint/parser',
+      plugins: ['@typescript-eslint/eslint-plugin'],
+      rules: {
+        '@typescript-eslint/no-unused-vars': [
+          'error',
+          {argsIgnorePattern: '^_'},
+        ],
+        'import/order': 'error',
+        'no-dupe-class-members': 'off',
+        'no-unused-vars': 'off',
+      },
+    },
     // to make it more suitable for running on code examples in docs/ folder
     {
       files: ['*.md'],
@@ -123,6 +136,7 @@ module.exports = {
           '^types/(.*)': './types/$1',
         },
       },
+      'eslint-import-resolver-typescript': true,
     },
   },
 };