+++ babel.config.js
@@ -2,6 +2,21 @@
 
 module.exports = {
   babelrcRoots: ['examples/*'],
+  overrides: [
+    {
+      presets: ['@babel/preset-flow'],
+      test: '**/*.js',
+    },
+    {
+      plugins: [
+        require.resolve(
+          './scripts/babel-plugin-jest-replace-ts-export-assignment.js'
+        ),
+      ],
+      presets: ['@babel/preset-typescript'],
+      test: /\.tsx?$/,
+    },
+  ],
   plugins: [
     ['@babel/plugin-transform-modules-commonjs', {allowTopLevelThis: true}],
     '@babel/plugin-transform-strict-mode',
@@ -14,6 +29,5 @@ module.exports = {
         targets: {node: 6},
       },
     ],
-    '@babel/preset-flow',
   ],
 };