+++ e2e/coverage-report/sum.js
@@ -7,6 +7,7 @@
 
 require('./sumDependency.js');
 require('./otherFile');
+require('./file');
 
 const uncoveredFunction = () => 1 + 'abc';