+++ packages/jest-circus/src/utils.js
@@ -22,12 +22,14 @@ import type {
   TestName,
   TestResults,
 } from 'types/Circus';
+// $FlowFixMe: Converted to TS
 import {convertDescriptorToString} from 'jest-util';
 import isGeneratorFn from 'is-generator-fn';
 import co from 'co';
 
 import StackUtils from 'stack-utils';
 
+// $FlowFixMe: Converted to TS
 import prettyFormat from 'pretty-format';
 
 import {getState} from './state';