+++ packages/jest-cli/src/reporters/coverage_worker.js
@@ -8,6 +8,7 @@
  */
 
 import type {GlobalConfig, ProjectConfig, Path} from 'types/Config';
+import type {CoverageReporterOptions} from './coverage_reporter';
 
 import exit from 'exit';
 import fs from 'fs';
@@ -18,6 +19,7 @@ export type CoverageWorkerData = {|
   globalConfig: GlobalConfig,
   config: ProjectConfig,
   path: Path,
+  options?: CoverageReporterOptions,
 |};
 
 export type {CoverageWorkerResult};
@@ -32,11 +34,13 @@ export function worker({
   config,
   globalConfig,
   path,
+  options,
 }: CoverageWorkerData): ?CoverageWorkerResult {
   return generateEmptyCoverage(
     fs.readFileSync(path, 'utf8'),
     path,
     globalConfig,
     config,
+    options && options.changedFiles,
   );
 }