+++ scripts/watch.js
@@ -13,6 +13,7 @@ const fs = require('fs');
 const {execSync} = require('child_process');
 const path = require('path');
 const chalk = require('chalk');
+const execa = require('execa');
 const getPackages = require('./getPackages');
 
 const BUILD_CMD = `node ${path.resolve(__dirname, './build.js')}`;
@@ -27,11 +28,12 @@ const exists = filename => {
 };
 const rebuild = filename => filesToBuild.set(filename, true);
 
-getPackages().forEach(p => {
+const packages = getPackages();
+packages.forEach(p => {
   const srcDir = path.resolve(p, 'src');
   try {
     fs.accessSync(srcDir, fs.F_OK);
-    fs.watch(path.resolve(p, 'src'), {recursive: true}, (event, filename) => {
+    fs.watch(srcDir, {recursive: true}, (event, filename) => {
       const filePath = path.resolve(srcDir, filename);
 
       if ((event === 'change' || event === 'rename') && exists(filePath)) {
@@ -54,6 +56,12 @@ getPackages().forEach(p => {
     // doesn't exist
   }
 });
+
+const packageWithTs = packages.filter(p =>
+  fs.existsSync(path.resolve(p, 'tsconfig.json'))
+);
+
+execa('tsc', ['-b', ...packageWithTs, '--watch'], {stdio: 'inherit'});
 
 setInterval(() => {
   const files = Array.from(filesToBuild.keys());